+++ docs/source/conf.py
@@ -24,9 +24,9 @@ copyright = '2019, NVIDIA'
 author = 'NVIDIA'
 
 # The short X.Y version
-version = '0.17'
+version = '0.18'
 # The full version, including alpha/beta/rc tags
-release = '0.17.0'
+release = '0.18.0'
 
 
 # -- General configuration ---------------------------------------------------